user state db

+++ src/app/commands/post.js
@@ -1,14 +1,21 @@
 //postear en un tablón desde el bot
 import { Markup } from "telegraf";
-import { clearUserState, getConfirmationMenu, getCachedPin } from "@/app/utils/utils";
-import { userStates, cache } from "@/app/utils/consts"
+import { getConfirmationMenu, getCachedPin } from "@/app/utils/utils";
+// import { userStates, cache } from "@/app/utils/consts"
 import { post } from "@/app/external/post";
+
+import { getUserState, updateUserState } from "../utils/userStates";
+import { replaceMenu, getMenu } from "../utils/menus";
+import { getComments, getPosts } from "./scan";
+
 
 
 
 export default async (ctx) => {
     ctx.deleteMessage().catch(() => { });
-    await ctx.reply(boardMenu.text, boardMenu.keyboard);
+    const sentMenu = await ctx.reply(boardMenu.text, boardMenu.keyboard);
+    const state = await updateUserState(String(ctx.from.id), { step: 'selecting_board' })
+    await replaceMenu(state.id, sentMenu.message_id, { type: 'dialog_menu' })
 }
 
 
@@ -26,9 +33,16 @@ const boardMenu = {
 
 
 export async function sendPost(ctx) {
-    const userId = ctx.from.id
-    const state = userStates[userId]
+    const userId = String(ctx.from.id)
+    // const state = userStates[userId]
+    const state = await getUserState(userId)
     const pin = await getCachedPin(userId)
+    const menu = await getMenu(state.id)
+
+    if (!state || !menu) {
+        return ctx.answerCbQuery("Sesión expirada o inválida.")
+    }
+
 
     if (!pin) {
         return ctx.answerCbQuery("Sesión expirada. Ingresa tu PIN nuevamente.")
@@ -40,17 +54,18 @@ export async function sendPost(ctx) {
 
     try {
 
-        const { board, content, menuMessageId } = state
+        const { boardId, content } = state
 
-        const { error, id, cont} = await post(content, board, pin, userId.toString()) 
+        const { error, id, cont } = await post(content, boardId, pin, userId.toString(), state.id)
 
         if (error) {
             console.log("Error en la API al postear: ", await response.text())
-            return ctx.telegram.editMessageText(ctx.chat.id, menuMessageId, null, "Error al publicar el post. Inténtalo más ahorita.", getConfirmationMenu(ctx.chat.id))
+            return ctx.telegram.editMessageText(ctx.chat.id, menu.messageId, null, "Error al publicar el post. Inténtalo más ahorita.", getConfirmationMenu(ctx.chat.id))
         }
 
-        await ctx.telegram.editMessageText(ctx.chat.id, menuMessageId, null, "Post enviado con éxito.")
-        clearUserState(userId);
+        await ctx.telegram.editMessageText(ctx.chat.id, menu.messageId, null, "Post enviado con éxito.")
+        // clearUserState(userId);
+
 
         const notif = await fetch(`${process.env.URL}/api/notify/post`, {
             method: 'POST',
@@ -61,9 +76,17 @@ export async function sendPost(ctx) {
             body: JSON.stringify({
                 content: cont,
                 id: id,
-                boardId: board
+                boardId: boardId
             })
         })
+
+        if (state.previousStep) {
+            const newState = await updateUserState(userId, { step: state.previousStep, previousStep: null })
+            const sentMsg = await ctx.reply('Volviendo al menú anterior...')
+            const newMenu = await replaceMenu(state.id, sentMsg.message_id, { type: 'waiting', currentPage: menu.currentPage })
+            newState.step === 'viewing_post' ? await getPosts(ctx, newMenu.currentPage) : await getComments(ctx, newMenu.currentPage)
+        }
+
     } catch (error) {
         console.log("Error: ", error)
     }
@@ -72,23 +95,39 @@ export async function sendPost(ctx) {
 export function setUpPostHandlers(bot) {
     bot.action(/post:(.*)/, async (ctx) => {
         try {
-            const board = ctx.match[1]
-            const userId = ctx.from.id
+            const boardId = ctx.match[1]
+            const userId = String(ctx.from.id)
+            const state = await getUserState(userId)
+            const menu = await getMenu(state.id)
+            if (!state || !menu) {
+                return ctx.answerCbQuery("Sesión expirada o inválida.")
+            }
 
 
-            const sentMenu = await ctx.editMessageText(`Envíame el texto para publicar en ${board}:`, Markup.inlineKeyboard([
+            const sentMenu = await ctx.editMessageText(`Envíame el texto para publicar en ${boardId}:`, Markup.inlineKeyboard([
                 [Markup.button.callback('volver', 'back'),
                 Markup.button.callback('cancelar', 'cancel')
 
                 ]
             ]))
-            const state = userStates[userId]
-            userStates[userId] = {
-                ...state,
+
+
+
+
+            await updateUserState(userId, {
                 step: 'waiting_text',
-                board: board,
-                menuMessageId: sentMenu.message_id
-            };
+                boardId: boardId,
+            })
+
+            await replaceMenu(state.id, sentMenu.message_id, { type: 'cancel_menu' })
+
+            // const state = userStates[userId]
+            // userStates[userId] = {
+            //     ...state,
+            //     step: 'waiting_text',
+            //     board: board,
+            //     menuMessageId: sentMenu.message_id
+            // };
 
             await ctx.answerCbQuery();
 
@@ -100,36 +139,60 @@ export function setUpPostHandlers(bot) {
 
 
     bot.action('back', async (ctx) => {
-        const userId = ctx.from.id;
-        const state = userStates[userId];
+        const userId = String(ctx.from.id);
+        // const state = userStates[userId];
+        const state = await getUserState(userId)
+
 
         if (state && state.step === 'waiting_confirmation') {
 
 
-            const sentMenu = await ctx.editMessageText(`Envíame el nuevo texto para /${state.board}/:`,
+            const sentMenu = await ctx.editMessageText(`Envíame el nuevo texto para /${state.boardId}/:`,
                 Markup.inlineKeyboard([
                     [Markup.button.callback('cancelar', 'cancel')]
                 ])
             );
 
-            userStates[userId] = {
-                ...state,
+            await updateUserState(userId, {
                 step: 'waiting_text',
-                menuMessageId: sentMenu.message_id,
+                boardId: state.boardId,
                 content: undefined
-            };
+            })
+
+            await replaceMenu(state.id, sentMenu.message_id, { type: 'cancel_menu' })
+
+            // userStates[userId] = {
+            //     ...state,
+            //     step: 'waiting_text',
+            //     menuMessageId: sentMenu.message_id,
+            //     content: undefined
+            // };
 
         } else {
 
-            clearUserState(userId);
+            // clearUserState(userId);
             await ctx.editMessageText(boardMenu.text, boardMenu.keyboard);
         }
         await ctx.answerCbQuery();
     })
 
-    bot.action('cancel', (ctx) => {
-        clearUserState(ctx.from.id);
+    bot.action('cancel', async (ctx) => {
+        // clearUserState(ctx.from.id);
+        const state = await getUserState(ctx.from.id.toString())
+        const menu = await getMenu(state.id)
+        if (!state || !menu) {
+            return ctx.answerCbQuery("Sesión expirada o inválida.")
+        }
+
         ctx.deleteMessage();
+
+        if (state.previousStep) {
+            const newState = await updateUserState(String(ctx.from.id), { step: state.previousStep, previousStep: null })
+            const sentMsg = await ctx.reply('Volviendo al menú anterior...')
+            const newMenu = await replaceMenu(state.id, sentMsg.message_id, { type: 'waiting', currentPage: menu.currentPage })
+            newState.step === 'viewing_post' ? await getPosts(ctx, newMenu.currentPage) : await getComments(ctx, newMenu.currentPage)
+        }
+
 
     });
 
@@ -137,29 +200,35 @@ export function setUpPostHandlers(bot) {
     bot.action('send', async (ctx) => {
 
         await sendPost(ctx)
-        
-    })  
 
-bot.action('pin', async (ctx) => {
-    const userId = ctx.from.id
-    const state = userStates[userId]
+    })
 
-    if (!state) {
-        await ctx.answerCbQuery("Sesión expirada.")
-        return
-    }
+    bot.action('pin', async (ctx) => {
+        const userId = String(ctx.from.id)
+        // const state = userStates[userId]
+        const state = await getUserState(userId)
 
-    const messageId = ctx.callbackQuery.message.message_id;
+        if (!state) {
+            await ctx.answerCbQuery("Sesión expirada.")
+            return
+        }
 
-    await ctx.editMessageText("Elige un PIN de 6 dígitos. Este PIN se usará para generar tu llave secreta. ⚠ Importante: No guardamos este PIN, si lo olvidas no podrás editar ni borrar tus mensajes. Anótalo o recuérdalo", Markup.inlineKeyboard([
-        [Markup.button.callback('Cancelar', 'cancel')]
-    ]))
+        const messageId = ctx.callbackQuery.message.message_id;
 
-    userStates[userId] = {
-        ...state,
-        step: 'waiting_pin',
-        menuMessageId: messageId
-    };
-    await ctx.answerCbQuery();
-})
+        await ctx.editMessageText("Elige un PIN de 6 dígitos. Este PIN se usará para generar tu llave secreta. ⚠ Importante: No guardamos este PIN, si lo olvidas no podrás editar ni borrar tus mensajes. Anótalo o recuérdalo", Markup.inlineKeyboard([
+            [Markup.button.callback('Cancelar', 'cancel')]
+        ]))
+
+        await updateUserState(userId, {
+            step: 'waiting_pin',
+        })
+        await replaceMenu(state.id, messageId, { type: 'cancel_menu' })
+
+        // userStates[userId] = {
+        //     ...state,
+        //     step: 'waiting_pin',
+        //     menuMessageId: messageId
+        // };
+        await ctx.answerCbQuery();
+    })
 }
